pyproject.toml
[project]
name = "agentic-ai"
version = "0.1.0"
description = "A comprehensive repository for building context-aware, multi-agent AI applications using LangGraph and modern LLM orchestration patterns."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langgraph>=0.5.1",
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-community>=0.3.0",
    "langchain-chroma>=0.1.0",
    "chromadb>=0.5.0",
    "openai>=1.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "tiktoken>=0.7.0",
    "numpy>=1.24.0",
    "faiss-cpu>=1.7.0",
    "sqlalchemy>=2.0.41",
    "langchain-postgres>=0.0.15",
    "psycopg2-binary>=2.9.10",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src"]
